Made code more modular.
