+++ frontend/multi-oracle-defi/src/hooks/useContractAddresses.js
@@ -1,0 +1,6 @@
+export const useContractAddresses = () => {
+  return {
+    lendingContractAddress: "0x1e88e0dc3924D9869A1Ed7d86197341dd459CF48",
+    CUSDContractAddress: "0x3d24dA1CB3C58C10DBF2Df035B3577624a88E63A",
+  };
+};
