+++ frontend/multi-oracle-defi/src/screens/LendingBorrowing.js
@@ -1,139 +1,41 @@
 import React, { useState } from "react";
-import {
-  useAccount,
-  useReadContract,
-  useWriteContract,
-  useWaitForTransactionReceipt,
-  useBalance,
-} from "wagmi";
-import { parseEther, formatEther } from "viem";
-import { Box, Typography, Container, Grid } from "@mui/material";
+import { Box, Container, Grid } from "@mui/material";
 import { AccountBalance, TrendingUp } from "@mui/icons-material";
-import LendingBorrowingABI from "../utils/LendingBorrowingabi.json";
-import CUSDABI from "../utils/CUSDabi.json";
+import { formatEther } from "viem";
 import { ActionPanel } from "../components/ActionPanel";
-import { InfoCard } from "../components/infoCard";
+import { InfoCard } from "../components/InfoCard";
+import { ProtocolHeader } from "../components/ProtocolHeader";
+import { useContractAddresses } from "../hooks/useContractAddresses";
+import { useProtocolData } from "../hooks/useProtocolData";
+import { useUserData } from "../hooks/useUserData";
+import { useContractActions } from "../hooks/useContractActions";
 
 const LendingBorrowing = () => {
-  const { address } = useAccount();
   const [activeTab, setActiveTab] = useState("deposit");
   const [amount, setAmount] = useState("");
 
-  // Contract addresses
-  const lendingContractAddress = "0x1e88e0dc3924D9869A1Ed7d86197341dd459CF48";
-  const CUSDContractAddress = "0x3d24dA1CB3C58C10DBF2Df035B3577624a88E63A";
-
-  // Contract reads
-  const { data: ethPrice } = useReadContract({
-    address: lendingContractAddress,
-    abi: LendingBorrowingABI,
-    functionName: "getLatestPrice",
-  });
-
-  const { data: maxBorrow } = useReadContract({
-    address: lendingContractAddress,
-    abi: LendingBorrowingABI,
-    functionName: "getMaxBorrowAmount",
-    account: address,
-  });
-
-  const { data: userData } = useReadContract({
-    address: lendingContractAddress,
-    abi: LendingBorrowingABI,
-    functionName: "users",
-    args: [address],
-  });
-
-  const { data: contractCUSDBalance } = useReadContract({
-    address: CUSDContractAddress,
-    abi: CUSDABI,
-    functionName: "balanceOf",
-    args: [lendingContractAddress],
-  });
-
-  const { data: userCUSDBalance } = useReadContract({
-    address: CUSDContractAddress,
-    abi: CUSDABI,
-    functionName: "balanceOf",
-    args: [address],
-  });
-
-  const { data: contractETHBalance } = useBalance({
-    address: lendingContractAddress,
-  });
-
-  // Contract writes
+  const { lendingContractAddress, CUSDContractAddress } =
+    useContractAddresses();
+  const { ethPrice, contractCUSDBalance, contractETHBalance } = useProtocolData(
+    lendingContractAddress,
+    CUSDContractAddress
+  );
+  const { maxBorrow, userData, userCUSDBalance } = useUserData(
+    lendingContractAddress,
+    CUSDContractAddress
+  );
   const {
-    writeContract,
-    data: writeTxData,
-    isLoading: isWriteLoading,
+    handleAction,
+    handleApprove,
+    writeTxData,
+    isWriteLoading,
+    isTxLoading,
     error,
-  } = useWriteContract();
-
-  // Transaction receipts
-  const { isLoading: isTxLoading } = useWaitForTransactionReceipt({
-    hash: writeTxData?.hash,
-  });
+  } = useContractActions(lendingContractAddress, CUSDContractAddress);
 
   const handleTabChange = (event, newValue) => {
     setActiveTab(newValue);
     setAmount("");
-  };
-
-  const handleAction = async (action) => {
-    if (!amount) return;
-
-    try {
-      switch (action) {
-        case "borrow":
-          await writeContract({
-            address: lendingContractAddress,
-            abi: LendingBorrowingABI,
-            functionName: "borrow",
-            args: [amount],
-          });
-          break;
-        case "repay":
-          await writeContract({
-            address: lendingContractAddress,
-            abi: LendingBorrowingABI,
-            functionName: "repay",
-            args: [amount],
-          });
-          break;
-        case "withdraw":
-          await writeContract({
-            address: lendingContractAddress,
-            abi: LendingBorrowingABI,
-            functionName: "withdrawCollateral",
-            args: [parseEther(amount)],
-          });
-          break;
-        default:
-          await writeContract({
-            address: lendingContractAddress,
-            abi: LendingBorrowingABI,
-            functionName: "depositCollateral",
-            value: parseEther(amount),
-          });
-          break;
-      }
-    } catch (error) {
-      console.error(`${action} error:`, error);
-    }
-  };
-
-  const handleApprove = async () => {
-    try {
-      await writeContract({
-        address: CUSDContractAddress,
-        abi: CUSDABI,
-        functionName: "approve",
-        args: [lendingContractAddress, 1000000],
-      });
-    } catch (error) {
-      console.error("Approve error:", error);
-    }
   };
 
   return (
@@ -145,17 +47,9 @@ const LendingBorrowing = () => {
       }}
     >
       <Container maxWidth="xl">
-        <Box textAlign="center" mb={6}>
-          <Typography variant="h3" component="h1" color="white" gutterBottom>
-            Lending & Borrowing Protocol
-          </Typography>
-          <Typography variant="h6" color="rgba(255, 255, 255, 0.7)">
-            Deposit ETH, Borrow CUSD, and more!
-          </Typography>
-        </Box>
+        <ProtocolHeader />
 
         <Grid container spacing={4}>
-          {/* Protocol Info Card */}
           <Grid item xs={12} md={4}>
             <InfoCard
               title="Protocol Stats"
@@ -183,12 +77,11 @@ const LendingBorrowing = () => {
             />
           </Grid>
 
-          {/* Action Card */}
           <Grid item xs={12} md={4}>
             <ActionPanel
               activeTab={activeTab}
               handleTabChange={handleTabChange}
-              handleAction={handleAction}
+              handleAction={(action) => handleAction(action, amount)}
               handleApprove={handleApprove}
               amount={amount}
               setAmount={setAmount}
@@ -200,7 +93,6 @@ const LendingBorrowing = () => {
             />
           </Grid>
 
-          {/* User Info Card */}
           <Grid item xs={12} md={4}>
             <InfoCard
               title="Your Position"
